trying to fix unit tests
