+++ packages/client-core/src/social/components/XruiNameplateComponent.ts
@@ -42,11 +42,11 @@ import {
   useEntityContext
 } from '@ir-engine/ecs'
 import { AvatarComponent } from '@ir-engine/engine/src/avatar/components/AvatarComponent'
-import { inFrustum } from '@ir-engine/engine/src/camera/functions/cameraFunctions'
 import { createUI } from '@ir-engine/engine/src/interaction/functions/createUI'
 import { getState } from '@ir-engine/hyperflux'
 import { NetworkObjectComponent } from '@ir-engine/network'
 import { TransformComponent } from '@ir-engine/spatial'
+import { inFrustum } from '@ir-engine/spatial/src/camera/functions/CameraFunctions'
 import { createTransitionState } from '@ir-engine/spatial/src/common/functions/createTransitionState'
 import { smootheLerpAlpha } from '@ir-engine/spatial/src/common/functions/MathLerpFunctions'
 import { EngineState } from '@ir-engine/spatial/src/EngineState'
